help: move to clap 4 and localize strings

+++ src/args.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::module_name_repetitions)]
 
-use crate::operations::SearchBy;
-use clap::{Parser, Subcommand, ValueHint};
+use crate::{fl, operations::SearchBy};
+use clap::{ArgAction, Parser, Subcommand, ValueHint};
 
 static VERSION: &str = concat!(
     env!("CARGO_PKG_VERSION"),
@@ -13,67 +13,54 @@ static VERSION: &str = concat!(
 #[derive(Debug, Clone, Parser)]
 #[clap(bin_name = "ame", name = "Amethyst", version = VERSION, about = env ! ("CARGO_PKG_DESCRIPTION"), infer_subcommands = true)]
 pub struct Args {
-    #[clap(subcommand)]
+    #[command(subcommand)]
     pub subcommand: Option<Operation>,
 
-    /// Sets the level of verbosity
-    #[clap(long, short, parse(from_occurrences), global = true)]
-    pub verbose: usize,
+    #[arg(long, short, action = ArgAction::Count, global = true, help = fl!("verbose"))]
+    pub verbose: u8,
 
-    /// Complete operation without prompting user
-    #[clap(long = "noconfirm", global = true)]
+    #[arg(long = "noconfirm", global = true, help = fl!("no-confirm"))]
     pub no_confirm: bool,
 
-    /// Make some commands have less output (only clean, upgrade, and install are supported)
-    #[clap(long, short, global = true)]
     // not gonna lie the only reason this option is here is because
     // i was trying to figure out if paccache had a --noconfirm option.
     // so basically, it doesn't, but hey, we might as well have it here
     // anyways as some pacman commands can have a --quiet flag passed
     // to them.
+    #[arg(long, short, global = true, help = fl!("quiet"))]
     pub quiet: bool,
 
-    /// Loops sudo in the background to ensure it doesn't time out during long builds
-    #[clap(long = "sudoloop", global = true)]
+    #[arg(long = "sudoloop", global = true, help = fl!("sudoloop"))]
     pub sudoloop: bool,
 
-    /// Sets a custom AUR clone and build directory for the specified operation
-    #[clap(long, short, global = true, value_hint = ValueHint::DirPath)]
+    #[arg(long, short, global = true, value_hint = ValueHint::DirPath, help = fl!("cachedir"))]
     pub cachedir: Option<String>,
 }
 
 #[derive(Debug, Clone, Subcommand)]
 pub enum Operation {
-    /// Installs or searches for a package in either the AUR or the Pacman-defined repositories
-    #[clap(bin_name = "ame", name = "sync", aliases = & [ "-S" ], visible_aliases = & ["install", "i"], short_flag = 'S')]
+    #[command(bin_name = "ame", name = "sync", aliases = & [ "-S" ], visible_aliases = & ["install", "i"], short_flag = 'S', about = fl!("install"))]
     Install(InstallArgs),
 
-    /// Removes a previously installed package
-    #[clap(bin_name = "ame", name = "remove", visible_aliases = & ["rm"], short_flag = 'R')]
+    #[command(bin_name = "ame", name = "remove", visible_aliases = & ["rm"], short_flag = 'R', about = fl!("remove"))]
     Remove(RemoveArgs),
 
-    /// Searches for packages matching a provided pattern in the AUR/repos [aliases: -Ss]
-    #[clap(bin_name = "ame", name = "search")]
+    #[command(bin_name = "ame", name = "search", about = fl!("search"))]
     Search(InstallArgs),
 
-    /// Queries installed packages
-    #[clap(bin_name = "ame", name = "query", short_flag = 'Q')]
+    #[command(bin_name = "ame", name = "query", short_flag = 'Q', about = fl!("query"))]
     Query(QueryArgs),
 
-    /// Upgrades locally installed packages to their latest versions (Default)
-    #[clap(bin_name = "ame", name = "upgrade", visible_aliases = & ["-Syu"])]
+    #[command(bin_name = "ame", name = "upgrade", visible_aliases = & ["-Syu"], about = fl!("upgrade"))]
     Upgrade(UpgradeArgs),
 
-    /// Generates shell completions for supported shells (bash, fish, elvish, pwsh)
-    #[clap(bin_name = "ame", name = "gencomp", short_flag = 'G')]
+    #[command(bin_name = "ame", name = "gencomp", short_flag = 'G', about = fl!("gencomp"))]
     GenComp(GenCompArgs),
 
-    /// Removes all orphaned packages
-    #[clap(bin_name = "ame", name = "clean", short_flag = 'C')]
+    #[command(bin_name = "ame", name = "clean", short_flag = 'C', about = fl!("clean"))]
     Clean,
 
-    /// Runs pacdiff
-    #[clap(bin_name = "ame", name = "diff", short_flag = 'd')]
+    #[command(bin_name = "ame", name = "diff", short_flag = 'd', about = fl!("diff"))]
     Diff,
 }
 
@@ -85,63 +72,51 @@ impl Default for Operation {
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct InstallArgs {
-    /// The name of the package(s) to install or search for
-    #[clap(required = true)]
+    #[arg(required = true, help = fl!("install-packages"))]
     pub packages: Vec<String>,
 
-    /// Operate only on AUR packages
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("install-aur"))]
     pub aur: bool,
 
-    /// Operate only on repo packages
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("install-repo"))]
     pub repo: bool,
 
-    /// Search packages for a given pattern instead of installing
-    #[clap(hidden = true, short = 's')]
+    #[arg(hide = true, short = 's', help = fl!("install-search"))]
     pub search: bool,
 
-    /// Searches by a specific field
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("install-by"))]
     pub by: Option<SearchBy>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct RemoveArgs {
-    /// The name of the package(s) to remove
-    #[clap(required = true)]
+    #[arg(required = true, help = fl!("remove-packages"))]
     pub packages: Vec<String>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct QueryArgs {
-    /// Lists AUR/foreign packages [-Qa, -Qm]
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("query-aur"))]
     pub aur: bool,
 
-    /// Lists repo/native packages [-Qr, -Qn]
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("query-repo"))]
     pub repo: bool,
 
-    /// Get information about a specific package
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("query-info"))]
     pub info: Option<String>,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct UpgradeArgs {
-    /// Upgrades only repo/native packages
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("upgrade-repo"))]
     pub repo: bool,
 
-    /// Upgrades only from the AUR
-    #[clap(long, short)]
+    #[arg(long, short, help = fl!("upgrade-aur"))]
     pub aur: bool,
 }
 
 #[derive(Default, Debug, Clone, Parser)]
 pub struct GenCompArgs {
-    /// The shell to generate completions for (bash, fish, elvish, pwsh, fig)
-    #[clap(required = true)]
+    #[arg(required = true, help = fl!("gencomp-shell"))]
     pub shell: String,
 }
